consolidate getting posts logic in getPosts function
remove pagination for simplicity of template
